+++ src/main/java/net/minecraft/client/SectionUpdateTracker.java
@@ -23,8 +23,7 @@ public class SectionUpdateTracker {
         );
     }
 
-    // MCRe NoiseFarlands: section 坐标 Long 化
-    public void setDirty(final long sectionX, final long sectionY, final long sectionZ, final boolean playerChanged) {
+    public void setDirty(final int sectionX, final int sectionY, final int sectionZ, final boolean playerChanged) {
         SectionUpdateTracker.SectionDirtyState section = this.storage.getValue(sectionX, sectionY, sectionZ);
         if (section != null) {
             section.setDirty(playerChanged);
